% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readgtf.R
\name{readgtf}
\alias{readgtf}
\title{Reads in a gtf file and builds a bed file with geneid and gene_name as columns 4 and 6 respectively}
\usage{
readgtf(gtf, filter = "gene")
}
\arguments{
\item{gtf}{path to a gtf file}

\item{filter}{filter to apply to column 3 of the gtf file}
}
\description{
Reads in a gtf file and builds a bed file with geneid and gene_name as columns 4 and 6 respectively
}
\examples{
readingtf = readgtf("/Users/dphansti/Tools/genomes/gencode.v19/gencode.v19.annotation.gtf_withproteinids")
head(readingtf)


}
